Holy bug fixes, updated for original pso2h

+++ src/packetdef.h
@@ -14,7 +14,7 @@ namespace PSO2DamageDump
 	class Packet 
     {
     public:
-        DWORD dataSize;
+        WORD dataSize;
         LPBYTE data;
 
         // Constructor takes LPBYTE*
